+++ fc/grails-app/services/DemoContestProcedureTurnService.groovy
@@ -15,7 +15,7 @@ class DemoContestProcedureTurnService
 	
 	static final String NOWIND = "0/0"
 	
-	int CreateTest(String testName, String printPrefix, boolean testExists)
+	long CreateTest(String testName, String printPrefix, boolean testExists)
 	{
 		fcService.printstart "Create test contest '$testName'"
